Feature/handle employees (#4)
* create admin members page and route

* wip

* create axios interceptors to handle jwt expiration use case

* create register user on company flow

* finished

* create user company helpers to translate user company roles. disable loged user rows actions on user company members datatable

* enable change session

* redesign users and company info component

+++ src/app/dashboard/services/admin/page.tsx
@@ -3,7 +3,12 @@ import { HeaderBreadcrumbItemProps } from "../../../../components/breadcrumbs/he
 import ButtonComponent from "../../../../components/buttons/component";
 import TextComponent from "../../../../components/inputs/text";
 import { IoSearchOutline } from "react-icons/io5";
-import { Modal, ModalBody, ModalContent, useDisclosure } from "@nextui-org/react";
+import {
+  Modal,
+  ModalBody,
+  ModalContent,
+  useDisclosure,
+} from "@nextui-org/react";
 import DatatableComponent from "../../../../components/datatable/component";
 import { DatatableColumnsProps } from "../../../../components/datatable/types";
 import { useContext, useEffect, useState } from "react";
@@ -15,8 +20,12 @@ import { StatusCodes } from "http-status-codes";
 import { PaginatedData } from "../../../../helpers/application_response/types";
 import { usePersistedStore } from "../../../../store/store";
 import LogoComponent from "../../../../components/logo/component";
-import useCreateServiceService, { CreateServiceProps } from "../../../../entities/service/services/create/use_create_service";
-import useUpdateServiceService, { UpdateServiceProps } from "../../../../entities/service/services/update/use_update_service";
+import useCreateServiceService, {
+  CreateServiceProps,
+} from "../../../../entities/service/services/create/use_create_service";
+import useUpdateServiceService, {
+  UpdateServiceProps,
+} from "../../../../entities/service/services/update/use_update_service";
 import useGetAllServicesService from "../../../../entities/service/services/get_all/use_get_all";
 import ServiceModel from "../../../../entities/service/model";
 import ServiceInfoForm from "../../../../entities/service/forms/info/component";
@@ -29,39 +38,38 @@ const HEADER_BREADCRUMBS: HeaderBreadcrumbItemProps[] = [
   },
   {
     text: "Dashboard",
-    url: "/dashboard"
+    url: "/dashboard",
   },
   {
     text: "Servicios del taller",
-    url: "/dashboard/services"
-  }
-]
+    url: "/dashboard/services",
+  },
+];
 
 const TABLE_COLUMNS: DatatableColumnsProps[] = [
   {
     key: "name",
-    label: "Nombre"
+    label: "Nombre",
   },
   {
     key: "description",
-    label: "Descripción"
+    label: "Descripción",
   },
   {
     key: "category",
-    label: "Categoria"
+    label: "Categoria",
   },
   {
     key: "price",
-    label: "Precio (REF)"
+    label: "Precio (REF)",
   },
   {
     key: "actions",
-    label: "Acciones"
-  }
+    label: "Acciones",
+  },
 ];
 
 function AdminServicesPage() {
-
   const { authReducer } = usePersistedStore();
 
   const { token, sessionType } = authReducer;
@@ -74,24 +82,29 @@ function AdminServicesPage() {
 
   const { dispatch: toasterDispatch } = useContext(ToasterContext);
 
-  const [ page, setPage ] = useState<number>(1);
+  const [page, setPage] = useState<number>(1);
 
-  const { isGettingServices, payloadState, performGetAllServices } = useGetAllServicesService();
+  const { isGettingServices, payloadState, performGetAllServices } =
+    useGetAllServicesService();
 
   const {
     isOpen: isCreateNewServiceFormOpen,
     onOpenChange: onCreateNewServiceFormOpenChange,
     onOpen: onCreateNewServiceFormOpen,
-    onClose: onCreateNewServiceFormClose
+    onClose: onCreateNewServiceFormClose,
   } = useDisclosure();
 
-  const { isCreatingService, payloadState: createServiceState, performCreateService } = useCreateServiceService();
+  const {
+    isCreatingService,
+    payloadState: createServiceState,
+    performCreateService,
+  } = useCreateServiceService();
 
   const {
     isOpen: isUpdateServiceFormOpen,
     onOpenChange: onUpdateServiceFormOpenChange,
     onOpen: onUpdateServiceFormOpen,
-    onClose: onUpdateServiceFormClose
+    onClose: onUpdateServiceFormClose,
   } = useDisclosure();
 
   const { isUpdatingService, performUpdateService } = useUpdateServiceService();
@@ -100,37 +113,44 @@ function AdminServicesPage() {
     isOpen: isDeleteServiceOpen,
     onOpenChange: onDeleteServiceOpenChange,
     onOpen: onDeleteServiceOpen,
-    onClose: onDeleteServiceClose
+    onClose: onDeleteServiceClose,
   } = useDisclosure();
 
-  const { isUpdatingService: isDeletingServiceLoading, performUpdateService: performDeleteService } = useUpdateServiceService();
+  const {
+    isUpdatingService: isDeletingServiceLoading,
+    performUpdateService: performDeleteService,
+  } = useUpdateServiceService();
 
   useEffect(() => {
     setBreadcrumbs(HEADER_BREADCRUMBS);
   }, []);
 
   useEffect(() => {
-    performGetAllServices({page, company_id: sessionType?.company_id!});
+    performGetAllServices({ page, company_id: sessionType?.company_id! });
   }, [page]);
 
-
   useEffect(() => {
-    console.log("coyelabadre", createServiceState);
-    if (createServiceState !== 'not loaded') {
-      if (createServiceState.status === StatusCodes.CREATED)  {
-        toasterDispatch({payload: 'Creación de servicio exitoso', type: 'SUCCESS'});
+    if (createServiceState !== "not loaded") {
+      if (createServiceState.status === StatusCodes.CREATED) {
+        toasterDispatch({
+          payload: "Creación de servicio exitoso",
+          type: "SUCCESS",
+        });
         setPage(1);
       } else {
-        toasterDispatch({payload: createServiceState.errorMessage || 'No se pudo crear el nuevo servicio', type: 'ERROR'});
+        toasterDispatch({
+          payload:
+            createServiceState.errorMessage ||
+            "No se pudo crear el nuevo servicio",
+          type: "ERROR",
+        });
       }
     }
   }, [createServiceState]);
 
   useEffect(() => {
-    if (datatableAction.action === 'update')
-      onUpdateServiceFormOpen();
-    if (datatableAction.action === 'delete')
-      onDeleteServiceOpen();
+    if (datatableAction.action === "update") onUpdateServiceFormOpen();
+    if (datatableAction.action === "delete") onDeleteServiceOpen();
   }, [datatableAction]);
 
   function controlCreateServiceResponse(data: CreateServiceProps) {
@@ -141,45 +161,64 @@ function AdminServicesPage() {
   }
 
   function handleCreateNewService(values: ServiceModel) {
-    performCreateService({service: values, token: token!, company_id: sessionType?.company_id!}, controlCreateServiceResponse);
+    performCreateService(
+      { service: values, token: token!, company_id: sessionType?.company_id! },
+      controlCreateServiceResponse
+    );
   }
 
   function controlUpdateServiceResponse(data: UpdateServiceProps) {
     if (data.status === StatusCodes.OK) {
       onUpdateServiceFormClose();
-      toasterDispatch({payload: "Servicio actualizado correctamente", type: "SUCCESS"});
+      toasterDispatch({
+        payload: "Servicio actualizado correctamente",
+        type: "SUCCESS",
+      });
       setPage(1);
       return;
     } else {
-      toasterDispatch({payload: data.errorMessage!, type: "ERROR"});
+      toasterDispatch({ payload: data.errorMessage!, type: "ERROR" });
     }
 
-    toasterDispatch({payload: "No se pudo actualizar el servicio", type: "ERROR"});
+    toasterDispatch({
+      payload: "No se pudo actualizar el servicio",
+      type: "ERROR",
+    });
   }
 
   function handleUpdateService(values: ServiceModel) {
-    performUpdateService({service: values, company_id: sessionType?.company_id!, token: token!}, controlUpdateServiceResponse);
+    performUpdateService(
+      { service: values, company_id: sessionType?.company_id!, token: token! },
+      controlUpdateServiceResponse
+    );
   }
 
   function controlDeleteServiceResponse(data: UpdateServiceProps) {
     if (data.status === StatusCodes.OK) {
       onDeleteServiceClose();
-      toasterDispatch({payload: "Servicio eliminado correctamente", type: "SUCCESS"});
+      toasterDispatch({
+        payload: "Servicio eliminado correctamente",
+        type: "SUCCESS",
+      });
       setPage(1);
       return;
     } else {
-      toasterDispatch({payload: data.errorMessage!, type: "ERROR"});
+      toasterDispatch({ payload: data.errorMessage!, type: "ERROR" });
     }
 
-    toasterDispatch({payload: "No se pudo eliminar el servicio", type: "ERROR"});
+    toasterDispatch({
+      payload: "No se pudo eliminar el servicio",
+      type: "ERROR",
+    });
   }
 
   function handleDeleteService() {
-    performDeleteService({
-      service: {id: datatableAction.id!, is_active: false} as ServiceModel,
-      company_id: sessionType?.company_id!,
-      token: token!
-    },
+    performDeleteService(
+      {
+        service: { id: datatableAction.id!, is_active: false } as ServiceModel,
+        company_id: sessionType?.company_id!,
+        token: token!,
+      },
       controlDeleteServiceResponse
     );
   }
@@ -197,10 +236,10 @@ function AdminServicesPage() {
       >
         <ModalContent>
           <ModalBody className="flex flex-col">
-            <span className="font-bold text-2xl font-inter">Formulario de nuevo servicio</span>
-            <ServiceInfoForm
-              onSubmit={handleCreateNewService}
-            >
+            <span className="font-bold text-2xl font-inter">
+              Formulario de nuevo servicio
+            </span>
+            <ServiceInfoForm onSubmit={handleCreateNewService}>
               <ButtonComponent
                 color="primary"
                 text="Confirmar"
@@ -223,26 +262,32 @@ function AdminServicesPage() {
       >
         <ModalContent>
           <ModalBody>
-            {
-              payloadState === "not loaded" ? null :
-              !datatableAction.id ? <span>Seleccione un servicio para modificación</span> :
-                <div className="flex flex-col gap-5">
-                  <span className="font-bold text-2xl font-inter">Formulario de modificación de servicio</span>
-                  <ServiceInfoForm
-                    onSubmit={handleUpdateService}
-                    initialValues={(payloadState.payload as PaginatedData<ServiceModel>).data.find( element => element.id === datatableAction.id )!}
-                    requiredFields={false}
-                  >
-                    <ButtonComponent
-                      color="primary"
-                      text="Confirmar"
-                      type="submit"
-                      variant="solid"
-                      isLoading={isUpdatingService}
-                    />
-                  </ServiceInfoForm>
-                </div>
-            }
+            {payloadState === "not loaded" ? null : !datatableAction.id ? (
+              <span>Seleccione un servicio para modificación</span>
+            ) : (
+              <div className="flex flex-col gap-5">
+                <span className="font-bold text-2xl font-inter">
+                  Formulario de modificación de servicio
+                </span>
+                <ServiceInfoForm
+                  onSubmit={handleUpdateService}
+                  initialValues={
+                    (
+                      payloadState.payload as PaginatedData<ServiceModel>
+                    ).data.find((element) => element.id === datatableAction.id)!
+                  }
+                  requiredFields={false}
+                >
+                  <ButtonComponent
+                    color="primary"
+                    text="Confirmar"
+                    type="submit"
+                    variant="solid"
+                    isLoading={isUpdatingService}
+                  />
+                </ServiceInfoForm>
+              </div>
+            )}
           </ModalBody>
         </ModalContent>
       </Modal>
@@ -257,29 +302,43 @@ function AdminServicesPage() {
       >
         <ModalContent>
           <ModalBody>
-            {
-              payloadState === "not loaded" ? null :
-              !datatableAction.id ? <span>Seleccione una categoria para eliminar</span> :
-                <div className="flex flex-col justify-center items-center gap-5">
-                  <LogoComponent />
-                  <span className="font-inter">Está seguro de eliminar la categoria <strong>{(payloadState.payload as PaginatedData<ServiceModel>).data.find( element => element.id === datatableAction.id )!.name}</strong>?</span>
-                  <ButtonComponent
-                    color="primary"
-                    text="Confirmar"
-                    type="button"
-                    onClick={handleDeleteService}
-                    variant="solid"
-                    isLoading={isDeletingServiceLoading}
-                  />
-                </div>
-            }
+            {payloadState === "not loaded" ? null : !datatableAction.id ? (
+              <span>Seleccione una categoria para eliminar</span>
+            ) : (
+              <div className="flex flex-col justify-center items-center gap-5">
+                <LogoComponent />
+                <span className="font-inter">
+                  Está seguro de eliminar la categoria{" "}
+                  <strong>
+                    {
+                      (
+                        payloadState.payload as PaginatedData<ServiceModel>
+                      ).data.find(
+                        (element) => element.id === datatableAction.id
+                      )!.name
+                    }
+                  </strong>
+                  ?
+                </span>
+                <ButtonComponent
+                  color="primary"
+                  text="Confirmar"
+                  type="button"
+                  onClick={handleDeleteService}
+                  variant="solid"
+                  isLoading={isDeletingServiceLoading}
+                />
+              </div>
+            )}
           </ModalBody>
         </ModalContent>
       </Modal>
-      <div className='flex flex-col'>
+      <div className="flex flex-col">
         <div className="w-full flex justify-between items-center m-auto px-10">
-          <span className="text-3xl font-bold font-inter">Servicios de la empresa</span>
-          <div className='w-auto flex items-center gap-5'>
+          <span className="text-3xl font-bold font-inter">
+            Servicios de la empresa
+          </span>
+          <div className="w-auto flex items-center gap-5">
             <TextComponent
               name="name"
               onChange={() => {}}
@@ -296,7 +355,7 @@ function AdminServicesPage() {
                 type="button"
                 variant="solid"
                 onClick={onCreateNewServiceFormOpen}
-                startContent={<IoMdAdd className='w-5 h-5'/>}
+                startContent={<IoMdAdd className="w-5 h-5" />}
               />
             </div>
           </div>
@@ -304,7 +363,19 @@ function AdminServicesPage() {
         <div className="px-10 mt-10">
           <DatatableComponent
             columns={TABLE_COLUMNS}
-            data={payloadState === 'not loaded' ? [] : (payloadState.payload as PaginatedData<ServiceModel>).data.map( element => ({id: element.id, name: element.name, description: element.description, category: element.category!.name, price: `${element.price.toFixed(2)} REF`}) ) || []}
+            data={
+              payloadState === "not loaded"
+                ? []
+                : (
+                    payloadState.payload as PaginatedData<ServiceModel>
+                  ).data.map((element) => ({
+                    id: element.id,
+                    name: element.name,
+                    description: element.description,
+                    category: element.category!.name,
+                    price: `${element.price.toFixed(2)} REF`,
+                  })) || []
+            }
             selectedData={selectedValues}
             setSelectedData={setSelectedValues}
             selectionMode="multiple"
@@ -315,21 +386,22 @@ function AdminServicesPage() {
             isRowDataEditable
             isRowDataDeletable
           />
-          {
-            payloadState === 'not loaded' ? null :
+          {payloadState === "not loaded" ? null : (
             <div className="w-full flex justify-end mt-5">
               <PaginationComponent
                 page={page}
-                pages={(payloadState.payload as PaginatedData<ServiceModel>).total_pages}
+                pages={
+                  (payloadState.payload as PaginatedData<ServiceModel>)
+                    .total_pages
+                }
                 setPage={setPage}
               />
             </div>
-          }
+          )}
         </div>
       </div>
     </div>
   );
 }
 
-export default AdminServicesPage
-
+export default AdminServicesPage;
